chore: isolate hooks/, store/ logic
